pyproject.toml
[project]
name = "dex-liquidity-anomaly-detection"
version = "0.1.0"
description = ""
authors = [{ name = "Kazunori Osaki", email = "claudioropez.k@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow-providers-snowflake (>=5.3.1,<6.0.0)",
    "dbt-duckdb (>=1.9.3,<2.0.0)",
    "dbt-bigquery (>=1.9.1,<2.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "bentoml (>=1.4.12,<2.0.0)",
    "jupyterlab (>=4.4.1,<5.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "snowflake-connector-python (>=3.15.0,<4.0.0)",
    "apache-airflow-providers-slack (>=9.0.5,<10.0.0)",
    "cloudpickle (>=2.2.1,<3.0)",
    "graphviz (>=0.20.3,<1.0)",
    "pyarrow (>=16.1.0,<17.0)",
    "streamlit (>=1.45.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "papermill (>=2.6.0,<3.0.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "japanize-matplotlib (>=1.1.3,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dex-liquidity-anomaly-detection"
version = "0.1.0"
packages = [{ include = "scripts" }, { include = "services" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.8"
jupytext = "^1.17.1"
nbstripout = "^0.8.1"
pre-commit = "^4.2.0"


[tool.ruff]
line-length = 120
